Feat
This repository hosts a weather application developed using modern web technologies such as React Native for the app framework, Expo for easier development and deployment, and Tailwind CSS for styling. The app fetches and displays current weather conditions and a 7-day forecast, providing a seamless and responsive user experience. It is a great example of integrating multiple technologies to create a functional and visually appealing mobile application.